+++ templates/admin/url-shortener/short-urls-page.php
@@ -34,7 +34,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 		<div class="wrap">
 			<h1 class="wp-heading-inline"><?php esc_html_e( 'Short URLs', 'ffcertificate' ); ?></h1>
 			<?php if ( \FreeFormCertificate\Core\Capabilities::current_user_can_admin_or( 'ffc_view_settings' ) ) : ?>
-				<a href="<?php echo esc_url( admin_url( 'edit.php?post_type=ffc_form&page=ffc-settings&tab=url_shortener' ) ); ?>" class="page-title-action">
+				<a href="<?php echo esc_url( admin_url( 'admin.php?page=ffc-settings&tab=url_shortener' ) ); ?>" class="page-title-action">
 					<?php esc_html_e( 'Settings', 'ffcertificate' ); ?>
 				</a>
 			<?php endif; ?>